フレーム部分を作っています

+++ StageView.cpp
@@ -28,12 +28,12 @@ void StageView::Draw()
 {
     Transform stagetrans;
     for (int z = 0; z < pStageData_->stageRange.z; z++) {
-        stagetrans.position_.y = z;
+        stagetrans.position_.y = (float)z;
         for (int x = 0; x < pStageData_->stageRange.x; x++) {
-            stagetrans.position_.x = x;
+            stagetrans.position_.x = (float)x;
             for (int y = 0; y < pStageData_->stageRange.y; y++) {
-                stagetrans.position_.z = y;
-                stagetrans.rotate_.x = (int)(pStageData_->sData[x][y][z].rot) * 90;
+                stagetrans.position_.z = (float)y;
+                stagetrans.rotate_.x = (float)((int)(pStageData_->sData[x][y][z].rot) * 90);
                 Model::SetTransform(hModel[(int)(pStageData_->sData[x][y][z].tag)], stagetrans);
                 Model::Draw(hModel[(int)(pStageData_->sData[x][y][z].tag)]);
